+++ homework/daemon/npm-debug.log
@@ -1,0 +1,44 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/bin/nodejs', '/usr/bin/npm', 'start' ]
+2 info using npm@3.5.2
+3 info using node@v4.2.6
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle node-simple-http-daemon@1.0.0~prestart: node-simple-http-daemon@1.0.0
+6 silly lifecycle node-simple-http-daemon@1.0.0~prestart: no script for prestart, continuing
+7 info lifecycle node-simple-http-daemon@1.0.0~start: node-simple-http-daemon@1.0.0
+8 verbose lifecycle node-simple-http-daemon@1.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle node-simple-http-daemon@1.0.0~start: PATH: /usr/share/npm/bin/node-gyp-bin:/home/sasha/PythonProjects/itea-advanced-python/homework/daemon/node_modules/.bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin
+10 verbose lifecycle node-simple-http-daemon@1.0.0~start: CWD: /home/sasha/PythonProjects/itea-advanced-python/homework/daemon
+11 silly lifecycle node-simple-http-daemon@1.0.0~start: Args: [ '-c', 'node lib/app.js' ]
+12 info lifecycle node-simple-http-daemon@1.0.0~start: Failed to exec start script
+13 verbose stack Error: node-simple-http-daemon@1.0.0 start: `node lib/app.js`
+13 verbose stack spawn ENOENT
+13 verbose stack     at ChildProcess.<anonymous> (/usr/share/npm/lib/utils/spawn.js:17:16)
+13 verbose stack     at emitTwo (events.js:87:13)
+13 verbose stack     at ChildProcess.emit (events.js:172:7)
+13 verbose stack     at maybeClose (internal/child_process.js:821:16)
+13 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:211:5)
+14 verbose pkgid node-simple-http-daemon@1.0.0
+15 verbose cwd /home/sasha/PythonProjects/itea-advanced-python/homework/daemon
+16 error Linux 4.4.0-78-generic
+17 error argv "/usr/bin/nodejs" "/usr/bin/npm" "start"
+18 error node v4.2.6
+19 error npm  v3.5.2
+20 error file sh
+21 error code ELIFECYCLE
+22 error errno ENOENT
+23 error syscall spawn
+24 error node-simple-http-daemon@1.0.0 start: `node lib/app.js`
+24 error spawn ENOENT
+25 error Failed at the node-simple-http-daemon@1.0.0 start script 'node lib/app.js'.
+25 error Make sure you have the latest version of node.js and npm installed.
+25 error If you do, this is most likely a problem with the node-simple-http-daemon package,
+25 error not with npm itself.
+25 error Tell the author that this fails on your system:
+25 error     node lib/app.js
+25 error You can get information on how to open an issue for this project with:
+25 error     npm bugs node-simple-http-daemon
+25 error Or if that isn't available, you can get their info via:
+25 error     npm owner ls node-simple-http-daemon
+25 error There is likely additional logging output above.
+26 verbose exit [ 1, true ]
