+++ homework/daemon/package.json
@@ -1,0 +1,15 @@
+{
+"name": "node-simple-http-daemon",
+"version": "1.0.0",
+"description": "Simple HTTP Daemon written in Node.js",
+"main": "lib/app.js",
+"scripts": {
+"test": "echo \"Error: no test specified\" && exit 1",
+"start": "node lib/app.js"
+},
+"author": "Oleksandr Shmigelskii",
+"license": "MIT",
+"dependencies": {
+"daemon": "~1.1.0"
+}
+}
