+++ homework/daemon/rain_log/RAIN.log
@@ -1,0 +1,2 @@
+*** 2017-06-04: Starting up RAIN.
+*** 2017-06-04: Starting up RAIN.
