+++ travel_platform/skyscanner_data/skyscanner_data.py
@@ -15,57 +15,41 @@ outbound_date = config.get('api_skyscanner', 'outbound_date')
 inbound_date = config.get('api_skyscanner', 'inbound_date')
 flights_cache_service = FlightsCache(api_key)
 
+#FIXME: why my *args not work?
 
-def cheapest_quotes():
-    result = flights_cache_service.get_cheapest_quotes(
-        market='UK',
-        currency='GBP',
-        locale='en-GB',
-        originplace='SIN-sky',
-        destinationplace='KUL-sky',
-        outbounddate='2017-07-15',
-        inbounddate='2017-07-16').parsed
+def cheapest_quotes(*args):
+    result = flights_cache_service.get_cheapest_quotes(args).parsed
     return result
 
 
-def cheapest_price_by_route():
-    result = flights_cache_service.get_cheapest_price_by_route(
-        market='UK',
-        currency='GBP',
-        locale='en-GB',
-        originplace='SIN-sky',
-        destinationplace='KUL-sky',
-        outbounddate='2017-07-15',
-        inbounddate='2017-07-16').parsed
+def cheapest_price_by_route(*args):
+    result = flights_cache_service.get_cheapest_price_by_route(args).parsed
     return result
 
 
-def cheapest_price_by_date():
-    result = flights_cache_service.get_cheapest_price_by_date(
-        market='UK',
-        currency='GBP',
-        locale='en-GB',
-        originplace='SIN-sky',
-        destinationplace='KUL-sky',
-        outbounddate='2017-07-15',
-        inbounddate='2017-07-16').parsed
+def cheapest_price_by_date(*args):
+    result = flights_cache_service.get_cheapest_price_by_date(args).parsed
     return result
 
 
-def grid_prices_by_date():
-    result = flights_cache_service.get_grid_prices_by_date(
-        market='UK',
-        currency='GBP',
-        locale='en-GB',
-        originplace='SIN-sky',
-        destinationplace='KUL-sky',
-        outbounddate='2017-07-15',
-        inbounddate='2017-07-16').parsed
+def grid_prices_by_date(*args):
+    result = flights_cache_service.get_grid_prices_by_date(args).parsed
     return result
 
-print cheapest_quotes()
-print cheapest_price_by_route()
-print cheapest_price_by_date()
-print grid_prices_by_date()
+print cheapest_quotes(market, currency, locale, origin_place,
+                      destination_place, outbound_date,
+                      inbound_date)
+
+print cheapest_price_by_route(market, currency, locale, origin_place,
+                      destination_place, outbound_date,
+                      inbound_date)
+
+print cheapest_price_by_date(market, currency, locale, origin_place,
+                      destination_place, outbound_date,
+                      inbound_date)
+
+print grid_prices_by_date(market, currency, locale, origin_place,
+                      destination_place, outbound_date,
+                      inbound_date)
 
 sys.path.pop()
